+++ Sources/AST/Environment/Environment+Memory.swift
@@ -1,0 +1,68 @@
+//
+//  Environment+Memory.swift
+//  AST
+//
+//  Created by Hails, Daniel R on 22/08/2018.
+//
+
+extension Environment {
+  /// The memory size of a type, in terms of number of memory slots it occupies.
+  public func size(of type: RawType) -> Int {
+    switch type {
+    case .basicType(.event): return 0 // Events do not use memory.
+    case .basicType(_): return 1
+    case .fixedSizeArrayType(let rawType, let elementCount): return size(of: rawType) * elementCount
+    case .arrayType(_): return 1
+    case .rangeType(_): return 0 // Ranges do not use memory
+    case .dictionaryType(_, _): return 1
+    case .inoutType(_): fatalError()
+    case .any: return 0
+    case .errorType: return 0
+
+    case .stdlibType(let type):
+      return types[type.rawValue]!.properties.reduce(0) { acc, element in
+        return acc + size(of: element.value.rawType)
+      }
+    case .userDefinedType(let identifier):
+      if isEnumDeclared(identifier),
+        case .enumCase(let enumCase) = types[identifier]!.properties.first!.value.property{
+        return size(of: enumCase.hiddenType.rawType)
+      }
+      return types[identifier]!.properties.reduce(0) { acc, element in
+        return acc + size(of: element.value.rawType)
+      }
+    }
+  }
+
+  /// The memory offset of a property in a type.
+  public func propertyOffset(for property: String, enclosingType: RawTypeIdentifier) -> Int? {
+
+    var offsetMap = [String: Int]()
+    var offset = 0
+
+    let rootType = types[enclosingType]!
+
+    for p in rootType.orderedProperties.prefix(while: { $0 != property }) {
+      offsetMap[p] = offset
+      let propertyType = rootType.properties[p]!.rawType
+      let propertySize = size(of: propertyType)
+
+      offset += propertySize
+    }
+
+    return offset
+  }
+}
+
+/// A table representing the memory offset of each property in a type.
+struct OffsetTable {
+  private var storage = [String: Int]()
+
+  func offset(for propertyName: String) -> Int? {
+    return storage[propertyName]
+  }
+
+  init(offsetMap: [String: Int]) {
+    storage = offsetMap
+  }
+}
